requirements.txt
pyaudio>=0.2.11
websocket-client>=1.6.0
pynput>=1.7.6
pyobjc-framework-Cocoa>=9.0
pyobjc-framework-ApplicationServices>=9.0
python-dotenv>=1.0.0 